# docs/openapi_schema.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Product Discovery Copilot",
    "description": "Backend API for AI Product Discovery Copilot.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/analyze-feedback": {
      "post": {
        "summary": "Analyze Feedback",
        "operationId": "analyze_feedback_analyze_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/analyze-feedback/start": {
      "post": {
        "summary": "Start Analyze Feedback",
        "operationId": "start_analyze_feedback_analyze_feedback_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/runs/latest/status": {
      "get": {
        "summary": "Get Latest Analysis Run Status",
        "operationId": "get_latest_analysis_run_status_runs_latest_status_get",
        "parameters": [
          {
            "name": "wait_seconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 40,
              "minimum": 0,
              "default": 0,
              "title": "Wait Seconds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/status": {
      "get": {
        "summary": "Get Analysis Run Status",
        "operationId": "get_analysis_run_status_runs__run_id__status_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "wait_seconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 40,
              "minimum": 0,
              "default": 0,
              "title": "Wait Seconds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/manifest": {
      "get": {
        "summary": "Get Run Manifest",
        "operationId": "get_run_manifest_runs__run_id__manifest_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactManifest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/artifact/{artifact_name}": {
      "get": {
        "summary": "Get Run Artifact",
        "operationId": "get_run_artifact_runs__run_id__artifact__artifact_name__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "artifact_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/final-report": {
      "post": {
        "summary": "Save Final Report",
        "operationId": "save_final_report_runs__run_id__final_report_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalReportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisTimeWindow": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Time window type, e.g. relative"
          },
          "value": {
            "type": "string",
            "title": "Value",
            "description": "Time window value, e.g. 12_months"
          }
        },
        "type": "object",
        "required": [
          "type",
          "value"
        ],
        "title": "AnalysisTimeWindow"
      },
      "AnalyzeFeedbackRequest": {
        "properties": {
          "product": {
            "type": "string",
            "minLength": 1,
            "title": "Product"
          },
          "research_scope": {
            "type": "string",
            "minLength": 1,
            "title": "Research Scope"
          },
          "research_goal": {
            "type": "string",
            "minLength": 1,
            "title": "Research Goal"
          },
          "analysis_time_window": {
            "$ref": "#/components/schemas/AnalysisTimeWindow"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "included_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Included Topics"
          },
          "excluded_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Excluded Topics"
          },
          "research_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Research Questions"
          },
          "success_criteria": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Success Criteria"
          },
          "max_runtime_seconds": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 60.0,
            "title": "Max Runtime Seconds",
            "default": 1800
          },
          "debug": {
            "type": "boolean",
            "title": "Debug",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "product",
          "research_scope",
          "research_goal",
          "analysis_time_window",
          "included_topics",
          "excluded_topics",
          "research_questions",
          "success_criteria"
        ],
        "title": "AnalyzeFeedbackRequest"
      },
      "AnalyzeFeedbackResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "locked_brief": {
            "type": "object",
            "title": "Locked Brief"
          },
          "source_summary": {
            "items": {
              "$ref": "#/components/schemas/SourceSummaryItem"
            },
            "type": "array",
            "title": "Source Summary"
          },
          "processing_summary": {
            "$ref": "#/components/schemas/ProcessingSummary"
          },
          "quality_diagnostics": {
            "$ref": "#/components/schemas/QualityDiagnostics"
          },
          "research_question_coverage": {
            "items": {
              "$ref": "#/components/schemas/ResearchQuestionCoverage"
            },
            "type": "array",
            "title": "Research Question Coverage"
          },
          "feedback_clusters": {
            "items": {
              "$ref": "#/components/schemas/ClusterItem"
            },
            "type": "array",
            "title": "Feedback Clusters"
          },
          "opportunities": {
            "items": {
              "$ref": "#/components/schemas/OpportunityItem"
            },
            "type": "array",
            "title": "Opportunities"
          },
          "evidence_backed_segments": {
            "items": {
              "$ref": "#/components/schemas/EvidenceBackedSegment"
            },
            "type": "array",
            "title": "Evidence Backed Segments"
          },
          "metrics": {
            "$ref": "#/components/schemas/MetricsPayload"
          },
          "charts_data": {
            "$ref": "#/components/schemas/ChartsDataPayload"
          },
          "representative_quotes": {
            "items": {
              "$ref": "#/components/schemas/QuoteItem"
            },
            "type": "array",
            "title": "Representative Quotes"
          },
          "compact_gpt_payload": {
            "$ref": "#/components/schemas/CompactGPTPayload"
          },
          "artifact_manifest": {
            "$ref": "#/components/schemas/ArtifactManifest"
          },
          "processing_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Processing Notes"
          },
          "source_limitations": {
            "items": {
              "$ref": "#/components/schemas/SourceLimitation"
            },
            "type": "array",
            "title": "Source Limitations"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "locked_brief",
          "source_summary",
          "processing_summary",
          "quality_diagnostics",
          "research_question_coverage",
          "feedback_clusters",
          "metrics",
          "charts_data",
          "representative_quotes",
          "compact_gpt_payload",
          "artifact_manifest",
          "processing_notes",
          "source_limitations",
          "warnings"
        ],
        "title": "AnalyzeFeedbackResponse"
      },
...
      "ArtifactManifest": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactItem"
            },
            "type": "array",
            "title": "Artifacts"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "artifacts"
        ],
        "title": "ArtifactManifest"
      },
...
      "ClusterItem": {
        "properties": {
          "cluster_id": {
            "type": "string",
            "title": "Cluster Id"
          },
          "cluster_name": {
            "type": "string",
            "title": "Cluster Name"
          },
          "cluster_summary": {
            "type": "string",
            "title": "Cluster Summary"
          },
          "cluster_tier": {
            "type": "string",
            "title": "Cluster Tier",
            "default": "tier_1"
          },
          "cluster_size": {
            "type": "integer",
            "title": "Cluster Size",
            "default": 0
          },
          "cluster_cohesion_score": {
            "type": "number",
            "title": "Cluster Cohesion Score",
            "default": 0.0
          },
          "frequency": {
            "type": "integer",
            "title": "Frequency"
          },
          "dominant_signal": {
            "type": "string",
            "title": "Dominant Signal",
            "default": "mixed"
          },
          "pain_point_evidence_count": {
            "type": "integer",
            "title": "Pain Point Evidence Count",
            "default": 0
          },
          "positive_validation_count": {
            "type": "integer",
            "title": "Positive Validation Count",
            "default": 0
          },
          "request_signal_count": {
            "type": "integer",
            "title": "Request Signal Count",
            "default": 0
          },
          "mixed_signal_flag": {
            "type": "boolean",
            "title": "Mixed Signal Flag",
            "default": false
          },
          "source_distribution": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Source Distribution"
          },
          "time_distribution": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Time Distribution"
          },
          "representative_quotes": {
            "items": {
              "$ref": "#/components/schemas/QuoteItem"
            },
            "type": "array",
            "title": "Representative Quotes"
          },
          "example_feedback_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Example Feedback Ids"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "mapped_research_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mapped Research Questions"
          },
          "mapped_success_criteria": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mapped Success Criteria"
          },
          "repeat_listening_cause_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Repeat Listening Cause Tags"
          },
          "relevance_score": {
            "type": "number",
            "title": "Relevance Score"
          }
        },
        "type": "object",
        "required": [
          "cluster_id",
          "cluster_name",
          "cluster_summary",
          "frequency",
          "source_distribution",
          "time_distribution",
          "representative_quotes",
          "example_feedback_ids",
          "keywords",
          "relevance_score"
        ],
        "title": "ClusterItem"
      },
...
      "ErrorResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "error"
          },
          "error_type": {
            "type": "string",
            "title": "Error Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          }
        },
        "type": "object",
        "required": [
          "error_type",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "FinalReportRequest": {
        "properties": {
          "markdown": {
            "type": "string",
            "title": "Markdown"
          }
        },
        "type": "object",
        "required": [
          "markdown"
        ],
        "title": "FinalReportRequest"
      },
      "FinalReportResponse": {
        "properties": {
          "artifact_name": {
            "type": "string",
            "title": "Artifact Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "artifact_name",
          "url"
        ],
        "title": "FinalReportResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProcessingSummary": {
        "properties": {
          "records_collected": {
            "type": "integer",
            "title": "Records Collected"
          },
          "records_after_cleaning": {
            "type": "integer",
            "title": "Records After Cleaning"
          },
          "records_relevant": {
            "type": "integer",
            "title": "Records Relevant"
          },
          "records_after_deduplication": {
            "type": "integer",
            "title": "Records After Deduplication"
          },
          "exact_duplicates_removed": {
            "type": "integer",
            "title": "Exact Duplicates Removed",
            "default": 0
          },
          "normalized_duplicates_removed": {
            "type": "integer",
            "title": "Normalized Duplicates Removed",
            "default": 0
          },
          "near_duplicates_removed": {
            "type": "integer",
            "title": "Near Duplicates Removed",
            "default": 0
          },
          "source_failures": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Failures"
          },
          "source_warning_codes": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Source Warning Codes"
          },
          "expanded_collection_applied": {
            "type": "boolean",
            "title": "Expanded Collection Applied",
            "default": false
          },
          "expanded_collection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expanded Collection Reason"
          }
        },
        "type": "object",
        "required": [
          "records_collected",
          "records_after_cleaning",
          "records_relevant",
          "records_after_deduplication"
        ],
        "title": "ProcessingSummary"
      },
      "QualityDiagnostics": {
        "properties": {
          "total_collected": {
            "type": "integer",
            "title": "Total Collected"
          },
          "in_window_records": {
            "type": "integer",
            "title": "In Window Records"
          },
          "out_of_window_records": {
            "type": "integer",
            "title": "Out Of Window Records"
          },
          "relevant_records": {
            "type": "integer",
            "title": "Relevant Records"
          },
          "relevant_rate": {
            "type": "number",
            "title": "Relevant Rate"
          },
          "dedupe_rate": {
            "type": "number",
            "title": "Dedupe Rate"
          },
          "cluster_count": {
            "type": "integer",
            "title": "Cluster Count"
          },
          "average_records_per_cluster": {
            "type": "number",
            "title": "Average Records Per Cluster"
          },
          "single_record_cluster_count": {
            "type": "integer",
            "title": "Single Record Cluster Count"
          },
          "source_contamination_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Contamination Warnings"
          },
          "time_window_violations": {
            "type": "integer",
            "title": "Time Window Violations",
            "default": 0
          },
          "expanded_collection_applied": {
            "type": "boolean",
            "title": "Expanded Collection Applied",
            "default": false
          },
          "expanded_collection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expanded Collection Reason"
          }
        },
        "type": "object",
        "required": [
          "total_collected",
          "in_window_records",
          "out_of_window_records",
          "relevant_records",
          "relevant_rate",
          "dedupe_rate",
          "cluster_count",
          "average_records_per_cluster",
          "single_record_cluster_count"
        ],
        "title": "QualityDiagnostics"
      },
...
      "ResearchQuestionCoverage": {
        "properties": {
          "question_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "evidence_strength": {
            "type": "string",
            "title": "Evidence Strength"
          },
          "relevant_cluster_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Relevant Cluster Ids"
          },
          "source_coverage": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Source Coverage"
          },
          "record_count": {
            "type": "integer",
            "title": "Record Count",
            "default": 0
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "evidence_gaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Gaps"
          }
        },
        "type": "object",
        "required": [
          "question",
          "evidence_strength",
          "summary"
        ],
        "title": "ResearchQuestionCoverage"
      },
      "RunStatusResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_stage": {
            "type": "string",
            "title": "Current Stage"
          },
          "progress_percent": {
            "type": "integer",
            "title": "Progress Percent",
            "default": 0
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "estimated_minutes_remaining": {
            "type": "integer",
            "title": "Estimated Minutes Remaining",
            "default": 0
          },
          "estimated_seconds_remaining": {
            "type": "integer",
            "title": "Estimated Seconds Remaining",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "manifest_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Url"
          },
          "artifact_manifest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtifactManifest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "current_stage",
          "started_at",
          "updated_at",
          "message"
        ],
        "title": "RunStatusResponse"
      },
...
      "SourceSummaryItem": {
        "properties": {
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "queries_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Queries Used"
          },
          "records_collected": {
            "type": "integer",
            "title": "Records Collected"
          },
          "records_relevant": {
            "type": "integer",
            "title": "Records Relevant"
          },
          "date_range": {
            "$ref": "#/components/schemas/SourceDateRange"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "source_name",
          "source_type",
          "queries_used",
          "records_collected",
          "records_relevant",
          "date_range",
          "notes"
        ],
        "title": "SourceSummaryItem"
      },
...
    }
  },
  "servers": [
    {
      "url": "https://productdiscoverycopilot-production.up.railway.app",
      "description": "Public HTTPS base URL of the deployed backend."
    }
  ]
}